:recycle: refactor: use adapter & strategy control
