+++ src/libs/next-auth/auth.config.ts
@@ -1,23 +1,32 @@
 import type { NextAuthConfig } from 'next-auth';
 
-import { authEnv } from '@/config/auth';
-
 import { ssoProviders } from './sso-providers';
+import { LobeNextAuthDbAdapter } from './adapter';
+import { getAuthConfig } from '@/config/auth';
+
+const {
+  NEXT_AUTH_DEBUG,
+  NEXT_AUTH_SECRET,
+  NEXT_AUTH_SSO_SESSION_STRATEGY,
+  NEXT_AUTH_SSO_PROVIDERS,
+  NEXT_PUBLIC_ENABLE_NEXT_AUTH
+} = getAuthConfig();
 
 export const initSSOProviders = () => {
-  return authEnv.NEXT_PUBLIC_ENABLE_NEXT_AUTH
-    ? authEnv.NEXT_AUTH_SSO_PROVIDERS.split(/[,，]/).map((provider) => {
-        const validProvider = ssoProviders.find((item) => item.id === provider.trim());
+  return NEXT_PUBLIC_ENABLE_NEXT_AUTH
+    ? NEXT_AUTH_SSO_PROVIDERS.split(/[,，]/).map((provider) => {
+      const validProvider = ssoProviders.find((item) => item.id === provider.trim());
 
-        if (validProvider) return validProvider.provider;
+      if (validProvider) return validProvider.provider;
 
-        throw new Error(`[NextAuth] provider ${provider} is not supported`);
-      })
+      throw new Error(`[NextAuth] provider ${provider} is not supported`);
+    })
     : [];
 };
 
 // Notice this is only an object, not a full Auth.js instance
 export default {
+  adapter: LobeNextAuthDbAdapter(),
   callbacks: {
     // Note: Data processing order of callback: authorize --> jwt --> session
     async jwt({ token, user }) {
@@ -39,12 +48,15 @@ export default {
       return session;
     },
   },
-  debug: authEnv.NEXT_AUTH_DEBUG,
+  debug: NEXT_AUTH_DEBUG,
   pages: {
     error: '/next-auth/error',
     signIn: '/next-auth/signin',
   },
   providers: initSSOProviders(),
-  secret: authEnv.NEXT_AUTH_SECRET,
+  secret: NEXT_AUTH_SECRET,
+  session: {
+    strategy: NEXT_AUTH_SSO_SESSION_STRATEGY,
+  },
   trustHost: process.env?.AUTH_TRUST_HOST ? process.env.AUTH_TRUST_HOST === 'true' : true,
 } satisfies NextAuthConfig;
